chore(llc): refactor useFeePresetFiatValues to use in both LWDM
