+++ .changeset/gentle-eyes-deliver.md
@@ -1,0 +1,7 @@
+---
+"ledger-live-desktop": minor
+"live-mobile": minor
+"@ledgerhq/live-common": minor
+---
+
+chore(llc): refactor useFeePresetFiatValues to use in both LWDM
