+++ libs/ledger-live-common/.unimportedrc.json
@@ -642,6 +642,7 @@
     "src/flows/send/customFees/hooks/useStableGasOptions.ts",
     "src/flows/send/customFees/utils/customFeeUtils.ts",
     "src/flows/send/effects/hooks/useFlowEffects.ts",
+    "src/flows/send/hooks/useFeePresetFiatValuesCore.ts",
     "src/flows/send/hooks/useSendFlowAccount.ts",
     "src/flows/send/hooks/useSendFlowAmountReviewCore.ts",
     "src/flows/send/hooks/useSendFlowBusinessLogic.ts",
@@ -655,6 +656,7 @@
     "src/flows/send/recipient/utils/shouldShowMatchedAddress.ts",
     "src/flows/send/types.ts",
     "src/flows/send/uiConfig.ts",
+    "src/flows/send/utils/feeEstimation.ts",
     "src/flows/send/utils.ts",
     "src/flows/wizard/types.ts",
     "src/genericAwarenessModal/buildCarousel.ts",
